+++ src/command_transaction.cpp
@@ -415,7 +415,7 @@ void Command::exportTransactions() {
 
     vector<JRecord> records;
 
-    for (int i = 0;i < results.size();i++) {
+    for (size_t i = 0;i < results.size();i++) {
         DBAccount account = results.at(i);
 
         JRecord r = account.getRecord();
@@ -436,7 +436,7 @@ void Command::exportTransactionsAsCSV() {
 
     out << DBTransaction::getCSVHeader();
 
-    for (int i = 0;i < results.size();i++) {
+    for (size_t i = 0;i < results.size();i++) {
         DBTransaction transaction = results.at(i);
 
         string record = transaction.getCSVRecord();
