+++ src/db_transaction.h
@@ -164,7 +164,7 @@ class DBTransaction : public DBPayment {
 
             os << getDeleteAllStatement() << endl;
 
-            for (int i = 0;i < results.size();i++) {
+            for (size_t i = 0;i < results.size();i++) {
                 os << results[i].getInsertStatementForRestore() << endl;
             }
